requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
mysql-connector-python
python-dotenv
groq